Correct bug when recursively creating directory
Before resursively creating a directory, an asynchronous call is made to test if the directory already exists or not. With concurrent calls to mkdir, a directory could pass the existing test then fail while calling fs.mkdir returning that the directory already exists as the directory have been created by another loop. It now verifies that calling fs.mkdir does not send an error telling that the directory already exists, which must not be considered as an error.