Update MongoDB database interface relative to MongoDB driver 2.0
Several modifications :
Validate that the database collection to reach can be fetched and handle the error if not. No validation were performed.
Return only the expected results from a query. Connection object and other arguments specific to MongoDB were returned.
Replace deprecated "insert" method by "insertMany".
Replace deprecated "update" method by "updateMany".
Add a filter argument to "removeProp" method to be able to add filters to select the documents. The locked filter is now externalized using this argument.
Log errors in EntityProvider if no callback is specified for a query.