Wrap dictionary's translations into a global property
Dictionaries of translations used to be merged between core and plugins, with this behaviour, each plugin could override translations describe by the core. To avoid collisions, it is no longer possible.
Core translations are now wrapped under a global property "CORE" of the translations object. Plugins are encouraged to do the same, for example "example" plugin translations will should be available under a global "EXAMPLE" property.
If this global property is missing, it will be automatically added.