Remove user access control from EntityController
Access control errors were used in EntityController but no user is associated to an entity, a user can be associated to a content entity but not an entity.
EntityController no longer checks for user access errors and ContentController now overrides most part of EntityController methods to check for user access errors.