Add grunt task to order Angular application files
AngularJS application files need to be loaded in
the correct order. Module definitions have to be
loaded first, then the other elements. In a
components based application css files of sub
component have to be loaded after css files of
parent component to guarantee that the sub
component can override style of its parent
component.

Grunt task ngDpTask generates a JSON file
containing the list of CSS and JavaScript files of
the AngularJS application in the correct order,
ready to be loaded in a brower.