Fix forbidden error on removeEntitiesAction
ContentController.removeEntitiesActions could
remove content entities partially due to
permissions. When two content entities were
specified to be removed and only one could be
removed by the actual user, one was removed and a
forbidden error was returned. It now removes both
content entities or none. If one of the content
entities specified can't be removed, a forbidden
error is returned an no entity is removed.