Add uniqueness to MultipartParser fields
MultipartParser is capable of sanitizing the file
name and make it unique if a file with the same
name exists in the destination folder. But when no
file with the same name already exists then the
default file name is kept. It is now possible to
force the uniqueness of the file name even if no
file with the same name already exists. For this
purpose a property "unique" has been added to the
field configuration. Setting "unique" to true will
generate a unique file name for each files of this
field, no matter what.