Fix fileSystem.extract for incomplete archives
fileSystem.extract was hanging when trying to
extract an incomplete archive, the extractor
doesn't finish after the "error" event has been
dispatched which was blocking the process until
the extract timeout.
It now executes callback with an error as soon as
the extractor has dispatched the error event.
There is no extract timeout anymore.