passport: add support for LDAP STARTTLS
LDAP STARTTLS secures LDAP connection using LDAP
standard protocol and port instead of LDAPS which
is not part of the standard and is deprecated.

Unfortunately it actually doesn't work with
passport-ldapauth due to sub-dependency
ldapauth-fork which doesn't wait for the
connection to be secured before trying to
authenticate (see Github issue:
https://github.com/vesse/node-ldapauth-fork/issues/65
for more details).