dependencies: upgrade socket.io
Library socket.io has been upgraded from 2.3.0 to
4.1.2. Note that the library has been made
compatible with socket.io clients in v2 by
activating the allowEIO3 option. This option might
not persist over the next releases of socket.io so
clients are encouraged to upgrade to version 4.
Also note that the signature of
SocketServer.listen has changed and now expects
the list of allowed origins as second parameter.